backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.22
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.2
todoist-api-python>=3.1.0
pytest>=8.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.24.0